feat: atualiza entidades, dtos e regras de negocio do fluxo de pedidos

+++ order-service/src/main/java/edu/mateus/catalogoprodutos/orderservice/OrderItemResponseDTO.java
@@ -1,8 +1,10 @@
 package edu.mateus.catalogoprodutos.orderservice;
 
+import java.math.BigDecimal;
+
 public record OrderItemResponseDTO(
     Long id,
     Integer quantity,
-    Double singlePrice
+    BigDecimal singlePrice
 ) {
 }
